temp

+++ mago3d-user/src/main/java/com/gaia3d/service/ConverterService.java
@@ -32,7 +32,7 @@ public interface ConverterService {
 	 * @param converterJob
 	 * @return
 	 */
-	public Long insertConverter(String userId, String checkIds, ConverterJob converterJob);
+	public int insertConverter(String userId, String checkIds, ConverterJob converterJob);
 	
 	/**
 	 * update
